feature(category): Add check to prevent category deletion if associated products exist

+++ src/categories/categories.service.ts
@@ -1,32 +1,47 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { BadRequestException, ConflictException, forwardRef, Inject, Injectable, NotFoundException } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
-import { Model } from 'mongoose';
+import { Model, Types } from 'mongoose';
 import { CreateCategoryDto } from './dto/create-category.dto';
 import { UpdateCategoryDto } from './dto/update-category.dto';
 import { Category } from './schemas/category.schema';
+import { ProductsService } from 'src/products/products.service';
+import { IdParamDto } from 'src/common/dto/IdParamDTO';
 
 @Injectable()
 export class CategoriesService {
-  constructor(@InjectModel(Category.name) private categoryModel: Model<Category>) {}
+  constructor(
+    @InjectModel(Category.name) private categoryModel: Model<Category>,
+    @Inject(forwardRef(() => ProductsService)) private productsService: ProductsService,
+  ) {}
 
   async create(createCategoryDto: CreateCategoryDto): Promise<Category> {
-    const category = new this.categoryModel(createCategoryDto);
-    return category.save();
+    try {
+      const category = new this.categoryModel(createCategoryDto);
+      return await category.save();
+    } catch (error) {
+      if (error.code === 11000) {
+        throw new ConflictException('Category with this name already exists');
+      }
+      throw error;
+    }
   }
 
   async findAll(): Promise<Category[]> {
     return this.categoryModel.find().exec();
   }
-
   async findOne(id: string): Promise<Category> {
-    const category = await this.categoryModel.findById(id).exec();
+    if (!Types.ObjectId.isValid(id)) {
+      throw new BadRequestException('Invalid ID format.');
+    }
+  
+    const category = await this.categoryModel.findById(new Types.ObjectId(id)).exec();
     if (!category) {
       throw new NotFoundException(`Category with ID ${id} not found`);
     }
     return category;
   }
 
-  async update(id: string, updateCategoryDto: UpdateCategoryDto): Promise<Category> {
+  async update(id: IdParamDto, updateCategoryDto: UpdateCategoryDto): Promise<Category> {
     const updatedCategory = await this.categoryModel.findByIdAndUpdate(id, updateCategoryDto, { new: true }).exec();
     if (!updatedCategory) {
       throw new NotFoundException(`Category with ID ${id} not found`);
@@ -35,6 +50,12 @@ export class CategoriesService {
   }
 
   async remove(id: string): Promise<void> {
+    const products = await this.productsService.findByCategoryId(id); // Supondo que você tenha um método em ProductsService para buscar produtos por categoria
+
+    if (products.length > 0) {
+      throw new ConflictException('It is not possible to delete the category because it is linked to one or more products.');
+    }
+
     const result = await this.categoryModel.findByIdAndDelete(id).exec();
     if (!result) {
       throw new NotFoundException(`Category with ID ${id} not found`);
